+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed handling of floor division (`//`) syntax when only Luau FFlag is enabled
+- Fixed missing space when table is inside of Luau interpolated string expression (`{{` is invalid syntax)
 
 ## [0.19.1] - 2023-11-15